% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{gender_guess}
\alias{gender_guess}
\title{Gender}
\usage{
gender_guess(names)
}
\arguments{
\item{names}{A vector of names}
}
\value{
A factor indicating whether names are m/f
}
\description{
Guess the gender of a person
}

